Implement donation creation and retrieval by post ID

+++ Dto/PostDtos/PostDto.cs
@@ -5,7 +5,7 @@ public class PostDto
     public required string Title { get; set; }
     public required string Content { get; set; }
     public string? MediaUrl { get; set; }
-    public int AmountGained { get; set; }
+    public decimal AmountGained { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
